requirements.txt
appdirs==1.4.3
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
cycler==0.10.0
Flask==0.12.2
functools32==3.2.3.post2
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
nba-py==0.1.1a2
numpy==1.13.1
packaging==16.8
pandas==0.20.3
pkg-resources==0.0.0
py-goldsberry==1.0.1
pylibmc==1.5.2
pyparsing==2.2.0
python-dateutil==2.6.1
python-nvd3==0.14.2
python-slugify==1.1.4
pytz==2017.2
requests==2.18.4
scikit-learn==0.19.0
scipy==0.19.1
six==1.10.0
subprocess32==3.2.7
Unidecode==0.4.21
urllib3==1.22
Werkzeug==0.12.2
